Released version 0.6.0

+++ src/wcmCommon.c
@@ -654,7 +654,7 @@ void xf86WcmEvent(WacomCommonPtr common, unsigned int channel,
 	ds = *pState;
 
 	/* timestamp the state for velocity and acceleration analysis */
-	ds.sample = GetTimeInMillis();
+	ds.sample = (int)GetTimeInMillis();
 
 	DBG(10, ErrorF("xf86WcmEvent: c=%d i=%d t=%d s=%p x=%d y=%d b=%d "
 		"p=%d rz=%d tx=%d ty=%d aw=%d rw=%d t=%d df=%d px=%d st=%d\n",
